runtime-cleanup: trim more webkit2gtk3 gstreamer deps
Per https://bugzilla.redhat.com/show_bug.cgi?id=2031213 ,
webkit2gtk3 recently added deps on gstreamer1-plugins-good and
gstreamer1-plugins-bad-free, which in turn pull in a bunch of
audio/video-related libraries. None of this is necessary for
the installer environment, where we only need webkit2gtk3 as
yelp's HTML rendering engine to display the installer help,
which has no audio or video, so we need to tear it all out.

Signed-off-by: Adam Williamson <awilliam@redhat.com>